Support Mindgames, fix missing Icons, moved talents into talents from spells
