+++ src/analysis/retail/priest/shadow/CombatLogParser.tsx
@@ -1,10 +1,4 @@
-import {
-  BoonOfTheAscended,
-  FaeGuardians,
-  Mindgames,
-  ShadowfiendNormalizer,
-  UnholyNova,
-} from 'analysis/retail/priest/shared';
+import { TwistOfFate, Mindgames, ShadowfiendNormalizer } from 'analysis/retail/priest/shared';
 import MainCombatLogParser from 'parser/core/CombatLogParser';
 import ArcaneTorrent from 'parser/shared/modules/racials/bloodelf/ArcaneTorrent';
 import Channeling from 'parser/shared/normalizers/Channeling';
@@ -24,17 +18,16 @@ import Voidform from './modules/features/Voidform';
 import InsanityTracker from './modules/resources/InsanityTracker';
 import InsanityUsage from './modules/resources/InsanityUsage';
 import DevouringPlague from './modules/spells/DevouringPlague';
-import Dispersion from './modules/spells/Dispersion';
+import Dispersion from './modules/talents/Dispersion';
 import Shadowfiend from './modules/spells/Shadowfiend';
 import ShadowWordDeath from './modules/spells/ShadowWordDeath';
 import ShadowWordPain from './modules/spells/ShadowWordPain';
-import VampiricEmbrace from './modules/spells/VampiricEmbrace';
+import VampiricEmbrace from './modules/talents/VampiricEmbrace';
 import VampiricTouch from './modules/spells/VampiricTouch';
 import AuspiciousSpirits from './modules/talents/AuspiciousSpirits';
 import DeathAndMadness from './modules/talents/DeathAndMadness';
 import ShadowCrash from './modules/talents/ShadowCrash';
 import DarkVoid from './modules/talents/DarkVoid';
-import TwistOfFate from 'analysis/retail/priest/shared/TwistOfFate';
 import UnfurlingDarkness from './modules/talents/UnfurlingDarkness';
 import VoidTorrent from './modules/talents/VoidTorrent';
 import MindSear from './modules/talents/MindSear';
@@ -85,10 +78,7 @@ class CombatLogParser extends MainCombatLogParser {
     shadowfiendNormalizer: ShadowfiendNormalizer,
 
     // covenants:
-    unholyNova: UnholyNova,
     mindgames: Mindgames,
-    boonOfTheAscended: BoonOfTheAscended,
-    faeGuardians: FaeGuardians,
 
     arcaneTorrent: [ArcaneTorrent, { active: false }] as const,
   };
